Fix a crash when `log.format: json` in Logstash.
The problem was that when `on_shutdown` is invoked, we were logging (on
purpose) the fact that a connection was shutdown. The exception object
was passed to the logger. The exception included a reference to the
socket (com.rabbitmq.client.impl.SocketFrameHandler) that was previously
open to RabbitMQ. When serializing this exception to JSON, Jackson is
calling `getTimeout`  on this object, which throws a SocketException
because the socket is closed.

The result was that our logger call was itself throwing an exception
because Jackson/JSON called a method that threw an exception. Oops!

For posterity, this is what the message logged like:

    2017-05-04 13:03:05,191 AMQP Connection 127.0.0.1:5672 ERROR java.net.SocketException: Socket is closed java.net.SocketException: Socket is closed

This error was reported to occur when Logstash was shutting down (thus
telling the rabbitmq output to terminate). Because it occurred during
shutdown, I do not believe this caused any measurable data flow
problems. It did cause alarm for users, though.

Originally filed at: https://github.com/elastic/logstash/issues/6906

Fixes #38